Separate Login component into separate presentational and container components
